// src/app/enums/buff.ts
export enum Buff {
  GiftOfTheWild = 'GiftOfTheWild',
  PaladinBlessing = 'PaladinBlessing',
  StrengthAgility = 'StrengthAgility',
  PercentAP = 'PercentAP',
  PhysCrit = 'PhysCrit',
  MeleeHaste = 'MeleeHaste',
  GlobalDamage = 'GlobalDamage',
  GlobalHaste = 'GlobalHaste',
  Intellect = 'Intellect',
  SpellPower = 'SpellPower',
  SpellCrit = 'SpellCrit',
  SpellHaste = 'SpellHaste',
  Spirit = 'Spirit',
  Stamina = 'Stamina',
  Health = 'Health',
  BlessingOfSanctuary = 'BlessingOfSanctuary',
  PassiveHealing = 'PassiveHealing',
  ManaReplenishment = 'ManaReplenishment',
}